explicit opensearch version in docker-compose.yml to be compatible with fylr
Opensearch 2.12.0 refuses to start with "No custom admin password found. Please provide a password via the environment variable OPENSEARCH_INITIAL_ADMIN_PASSWORD". But if you set a strong password, there seems to be no method yet to configure this password with fylr. So we advise to stick with opensearch:2.11.1 for the time being.